[Fix] Fix issue #2
Related issue #2

Fix the problem with spaces in the filename.
If there are more than one space in the uploaded filename, the problem
was emerged when axios received the response from the api server. It
was not able to handle the spaces in utf-8 character, like
`filename*=utf-8''sample%20before_d2l.png`
To solve this problem, copy the File.name and replace the spaces using
regex pattern. Then, use this processed filename for result.